refactor: complete config system simplification
- Remove all migration and compatibility code (not needed per requirements)
- Delete backup files (.bak) from previous refactoring attempts
- Fix integration tests to work with new config system
- Implement proper merging behavior for security fields:
  - Most arrays replace as requested
  - Security fields (protected_paths, force_home, pack.ignore) merge for safety
- Reduce config package from ~2500 to 769 lines of code

The config system now uses koanf directly without transformations between
user and internal formats, making it much simpler and more maintainable.
